feat: SpEL 기반 권한 고도화 (본인 리소스만 접근 가능하도록 구현)

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicMessageService.java
@@ -26,6 +26,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
+import org.springframework.security.access.prepost.PostAuthorize;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -104,6 +106,7 @@ public class BasicMessageService implements MessageService {
 
   @Override
   @Transactional
+  @PostAuthorize("returnObject.author.id == principal.userDto.id")
   public MessageDto update(UUID id, MessageUpdateRequest dto) {
     log.debug("update 시작 - 입력값: {}", dto);
 
@@ -118,16 +121,20 @@ public class BasicMessageService implements MessageService {
 
   @Override
   @Transactional
+  @PreAuthorize("@messageSecurity.isAuthor(#id, principal.userDto.id)")
   public void delete(UUID id) {
     log.debug("delete 시작 - 입력값: {}", id);
 
-    Message message = messageRepository.findById(id)
-        .orElseThrow(() -> new MessageNotFoundException(id));
+    // 메시지가 없을 경우 예외를 던지기 보다 멱등성을 보장하는 방향으로 구현
+    var messageOptional = messageRepository.findById(id);
 
-    if (message.getAttachments() != null && !message.getAttachments().isEmpty()) {
-      binaryContentRepository.deleteAll(message.getAttachments());
+    if (messageOptional.isPresent()) {
+      Message message = messageOptional.get();
+      if (message.getAttachments() != null && !message.getAttachments().isEmpty()) {
+        binaryContentRepository.deleteAll(message.getAttachments());
+      }
+      messageRepository.delete(message);
+      log.info("메시지 삭제 완료 - messageId: {}", message.getId());
     }
-    messageRepository.delete(message);
-    log.info("메시지 삭제 완료 - messageId: {}", message.getId());
   }
 }
